Fix run-with
