+++ grevling/__init__.py
@@ -1,16 +1,19 @@
 from contextlib import contextmanager
 from difflib import Differ
+from enum import Enum
+import json
 import os
 from pathlib import Path
 import pydoc
 import shutil
 
-from typing import List, Iterable
+from typing import List, Iterable, Optional
 
 from fasteners import InterProcessLock
 import pandas as pd
 from typing_inspect import get_origin, get_args
 
+from .capture import ResultCollector
 from .plotting import Plot
 from .render import render
 from .schema import load_and_validate
@@ -19,7 +22,6 @@ from .filemap import FileMap
 from .script import ScriptTemplate
 from . import util, api
 from .runner import local as runner
-from .instance import Instance
 
 
 __version__ = '1.2.1'
@@ -152,9 +154,8 @@ class Case:
         df.to_parquet(self.dataframepath, engine='pyarrow', index=True)
 
     def has_data(self):
-        from instance import Status
         with self.lock():
-            for instance in self.instances(Status.Downloaded):
+            for _ in self.instances(Status.Downloaded):
                 return True
         return False
 
@@ -221,13 +222,24 @@ class Case:
 
         return True
 
-    def create_instances(self) -> Iterable[Instance]:
+    def create_instances(self) -> Iterable['Instance']:
         for i, ctx in enumerate(self.context_mgr.fullspace()):
             ctx['_index'] = i
             ctx['_logdir'] = render(self._logdir, ctx)
             yield Instance.create(self, ctx)
 
-    def instances(self, *statuses) -> Iterable[Instance]:
+    def create_instance(self, ctx: api.Context, logdir: Optional[Path] = None, index: Optional[int] = None) -> 'Instance':
+        ctx = self.context_mgr.evaluate_context(ctx)
+        if index is None:
+            index = 0
+        ctx['_index'] = index
+        if logdir is None:
+            logdir = render(self._logdir, ctx)
+        ctx['_logdir'] = str(logdir)
+        workspace = runner.LocalWorkspace(Path(ctx['_logdir']))
+        return Instance.create(self, ctx, local=workspace)
+
+    def instances(self, *statuses) -> Iterable['Instance']:
         for name in self.storage_spaces.workspace_names():
             instance = Instance(self, logdir=name)
             if statuses and instance.status not in statuses:
@@ -235,13 +247,11 @@ class Case:
             yield instance
 
     def capture(self):
-        from .instance import Status
         with self.lock():
             for instance in self.instances(Status.Downloaded):
                 instance.capture()
 
     def collect(self):
-        from .instance import Status
         with self.lock():
             data = self.load_dataframe()
             for instance in self.instances(Status.Downloaded):
@@ -254,8 +264,149 @@ class Case:
         for plot in self._plots:
             plot.generate_all(self)
 
-    @util.with_context('instance {index}')
-    def run_single(self, index, namespace, workspace, log_ws):
-        util.log.info(', '.join(f'{k}={repr(namespace[k])}' for k in self.parameters))
-        success = self.script.run(namespace, workspace.root, log_ws)
-        return success
+
+class Status(Enum):
+
+    Created = 'created'
+    Prepared = 'prepared'
+    Started = 'started'
+    Finished = 'finished'
+    Downloaded = 'downloaded'
+
+
+class Instance:
+
+    local: api.Workspace
+    local_book: api.Workspace
+
+    remote: Optional[api.Workspace]
+    remote_book: Optional[api.Workspace]
+
+    logdir: str
+    status: Status
+
+    _context: Optional[api.Context]
+    _status: Optional[Status]
+
+    @classmethod
+    def create(cls, case, context: api.Context, local = None) -> 'Instance':
+        obj = cls(case, context=context, local=local)
+        obj.status = Status.Created
+        obj.write_context()
+        return obj
+
+    def __init__(self, case, context: api.Context = None, logdir = None, local = None):
+        self._case = case
+        self._context = context
+
+        if context:
+            self.logdir = context['_logdir']
+        else:
+            self.logdir = logdir
+
+        if local is None:
+            self.local = self.open_workspace(case.storage_spaces)
+        else:
+            self.local = local
+
+        self.local_book = self.local.subspace('.grevling')
+        self.remote = self.remote_book = None
+        self._status = None
+
+    @property
+    def status(self):
+        if not self._status:
+            with self.local_book.open_file('status.txt', 'r') as f:
+                status = f.read()
+            self._status = Status(status)
+        return self._status
+
+    @status.setter
+    def status(self, value):
+        with self.local_book.open_file('status.txt', 'w') as f:
+            f.write(value.value)
+        self._status = value
+
+    @property
+    def context(self):
+        if self._context is None:
+            with self.local_book.open_file('context.json', 'r') as f:
+                self._context = json.load(f)
+        return self._context
+
+    @property
+    def types(self):
+        return self._case.types
+
+    def __getitem__(self, key):
+        return self.context[key]
+
+    def __setitem__(self, key, value):
+        self.context[key] = util.coerce(self.types[key], value)
+
+    @contextmanager
+    def bind_remote(self, spaces: api.WorkspaceCollection):
+        self.remote = self.open_workspace(spaces, 'WRK')
+        self.remote_book = self.remote.subspace('.grevling')
+        try:
+            yield
+        finally:
+            self.remote = self.remote_book = None
+
+    @property
+    def index(self):
+        return self.context['_index']
+
+    @property
+    def script(self):
+        return self._case.script.render(self.context)
+
+    def write_context(self):
+        with self.local_book.open_file('context.json', 'w') as f:
+            json.dump(self.context, f, sort_keys=True, indent=4, cls=util.JSONEncoder)
+
+    def open_workspace(self, workspaces, name=''):
+        return workspaces.open_workspace(self.logdir, name)
+
+    def prepare(self):
+        assert self.remote
+        assert self.status == Status.Created
+
+        src = self._case.local_space
+        util.log.debug(f"Using SRC='{src}', WRK='{self.remote}'")
+        self._case.premap.copy(self.context, src, self.remote, ignore_missing=self._case._ignore_missing)
+
+        self.status = Status.Prepared
+
+    def download(self):
+        assert self.remote
+        assert self.remote_book
+        assert self.status == Status.Finished
+
+        collector = ResultCollector(self.types)
+        collector.collect_from_dict(self.context)
+
+        self.remote_book.copy_all_to(self.local_book)
+        collector.collect_from_info(self.local_book)
+
+        ignore_missing = self._case._ignore_missing or not collector['_success']
+        self._case.postmap.copy(self.context, self.remote, self.local, ignore_missing=ignore_missing)
+
+        self._case.script.capture(collector, self.local_book)
+        collector.commit_to_file(self.local_book)
+
+        self.status = Status.Downloaded
+
+    def capture(self):
+        assert self.status == Status.Downloaded
+
+        collector = ResultCollector(self.types)
+        collector.collect_from_dict(self.context)
+        collector.collect_from_info(self.local_book)
+        self._case.script.capture(collector, self.local_book)
+        collector.commit_to_file(self.local_book)
+
+    def cached_capture(self):
+        collector = ResultCollector(self.types)
+        collector.collect_from_cache(self.local_book)
+        return collector
